+++ lib/src/nodes/changes/node_change.dart
@@ -45,4 +45,8 @@ abstract class NodeChange {
   @override
   @mustBeOverridden
   int get hashCode;
+
+  @mustBeOverridden
+  @override
+  String toString();
 }