+++ lib/src/nodes/changes/node_deletion.dart
@@ -62,4 +62,15 @@ class NodeDeletion extends NodeChange {
           oldState,
         ],
       );
+
+  @override
+  String toString() {
+    return 'NodeDeletion('
+        'originalPosition: $originalPosition, '
+        'into: $inNode, '
+        'oldState (direct owner): $oldState, '
+        'newState (direct owner): $newState, '
+        'sourceOwner: $sourceOwner, '
+        ')';
+  }
 }