+++ lib/src/nodes/changes/node_update.dart
@@ -34,4 +34,12 @@ class NodeUpdate extends NodeChange {
           oldState,
         ],
       );
+
+  @override
+  String toString() {
+    return 'NodeUpdate('
+        'OldState: $oldState, '
+        'newState: $newState, '
+        ')';
+  }
 }